Moved text field jsx file into folder
Left the styling in the file - couldn't get it to work with useStyles in styles.css!
